fix: harden TDD loop state and coverage handling
- store coverage as float64 to avoid precision loss in gate checks
- persist task ID sequence to prevent ID reuse after task removal
- reset per-task execution state when advancing to the next task
- keep RED phase active until tests are actually written
- structure refactor notes with file, suggestion, and rationale
- apply spec TDD default to tasks added during refinement

+++ internal/engine/loop/persist_coverage.go
@@ -8,7 +8,7 @@ func persistCoverage(c *engine.Context, report StageReport) error {
 		return err
 	}
 	if tr.Coverage == nil {
-		tr.Coverage = map[string]int{}
+		tr.Coverage = map[string]float64{}
 	}
 	for _, entry := range report.FileCoverage {
 		tr.Coverage[entry.File] = entry.Coverage
