+++ internal/engine/loop/red_coverage_feedback.go
@@ -13,7 +13,7 @@ func appendCoverageFeedback(b *strings.Builder, ctx ExecCtx) {
 	if len(ctx.State.LastCoverage) == 0 {
 		return
 	}
-	threshold := ctx.Settings.MinTestCoverage
+	threshold := float64(ctx.Settings.MinTestCoverage)
 	var lowFiles []string
 	for file, pct := range ctx.State.LastCoverage {
 		if pct < threshold {
@@ -27,7 +27,7 @@ func appendCoverageFeedback(b *strings.Builder, ctx ExecCtx) {
 	b.WriteString("\nThe following files have low test coverage and need additional tests:\n")
 	for _, file := range lowFiles {
 		pct := ctx.State.LastCoverage[file]
-		b.WriteString(fmt.Sprintf("- %s: %d%% < %d%%\n", file, pct, threshold))
+		b.WriteString(fmt.Sprintf("- %s: %.1f%% < %.0f%%\n", file, pct, threshold))
 	}
 	b.WriteString("Add tests to bring these files above the coverage threshold.\n")
 }
